requirements.txt

pymitter==0.3.0
gpiozero==1.5.1
pyserial==3.5
Flask==1.1.2
perfume==0.2
RPi.GPIO==0.7.0
Adafruit_ADS1x15==1.0.2
adafruit-circuitpython-servokit==1.3.2
Flask-SocketIO==5.0.1